apps/backends/stocks-backend/web-clients/terrapin/pyproject.toml
[tool.poetry]
name = "FMP API Сlient"
version = "0.1.0"
description = "API Client for site.financialmodelingprep.com"
authors = ["Васин Денис <vassin.denis.igorevich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
requests = "^2.32.3"
pydantic-settings = "^2.7.1"
aiohttp = "^3.11.12"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
pip-tools = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.flake8]
max-line-length = 79
